prov/cxi: Increase MR key size to 4 bytes
This is to support applications which may have been using an RKEY
outside the 2 byte range when the CXI provider was not
enforcing/checking the RKEY range.

NETCASSINI-3773

+++ prov/cxi/include/cxip.h
@@ -169,7 +169,7 @@
 /* 16 bits of MR keys equals the hardware MR resource limit. Could explore
  * increasing key range if needed.
  */
-#define CXIP_MR_KEY_SIZE sizeof(uint16_t)
+#define CXIP_MR_KEY_SIZE sizeof(uint32_t)
 #define CXIP_MR_KEY_MASK ((1ULL << (8 * CXIP_MR_KEY_SIZE)) - 1)
 
 static inline bool cxip_is_valid_mr_key(uint64_t key)
